smallShop/src/test/java/wwhstudycleanarchproject/smallShop/Member/adapter/in/web/SessionControllerTest.java
package wwhstudycleanarchproject.smallShop.Member.adapter.in.web;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import wwhstudycleanarchproject.smallShop.Member.adapter.in.web.session.SessionController;
import wwhstudycleanarchproject.smallShop.product.application.service.AuthenticationService;

import javax.security.auth.login.LoginException;

import static org.hamcrest.Matchers.containsString;
import static org.mockito.BDDMockito.given;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@WebMvcTest(SessionController.class)
class SessionControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private AuthenticationService authenticationService;

    @BeforeEach
    void setUp() {
        given(authenticationService.login(
                "tester@example.com", "test"))
                .willReturn("a.b.c.");

        given(authenticationService.login("", ""))
                .willThrow(new LoginException());
    }

    @Test
    void loginTestWithRightEmailAndPassword() throws Exception {
        mockMvc.perform(post("/session")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content("{\"email\":\"tester@example.com\"," +
                                "\"password\":\"test\"}"))
                .andExpect(status().isCreated())
                .andExpect(content().string(containsString(".")));
    }

    @Test
    void loginTestWithWrongEmailAndPassword() throws Exception {
        mockMvc.perform(post("/session")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content("{\"email\":\"tester@example.com\"," +
                                "\"password\":\"test\"}"))
                .andExpect(status().isBadRequest());
    }

}
